Refactor code after integration CollectionView
