+++ TaskBook/Core/ServiceLayer/ModuleBuilder.swift
@@ -10,8 +10,6 @@ import Foundation
 // MARK: - Buildable
 protocol Buildable {
     func buildContainerModule() -> ContainerViewController
-//    func buildCalendarModule() -> CalendarViewCell
-//    func buildTaskModule() -> TaskViewCell
 }
 
 // MARK: - ModuleBuilder
@@ -34,21 +32,4 @@ extension ModuleBuilder: Buildable {
         presenter.viewController = viewController
         return viewController
     }
-    
-//    func buildCalendarModule() -> CalendarViewCell {
-//        let viewController = CalendarViewCell()
-//        let presenter = CalendarPresenter(calendarHelper: calendarHelper,
-//                                          moduleBuilder: self)
-//        viewController.presenter = presenter
-//        presenter.viewController = viewController
-//        return viewController
-//    }
-//
-//    func buildTaskModule() -> TaskViewCell {
-//        let viewController = TaskViewCell()
-//        let presenter = TaskPresenter(moduleBuilder: self)
-//        viewController.presenter = presenter
-//        presenter.viewController = viewController
-//        return viewController
-//    }
 }
